backend/requirements.txt
six~=1.16.0
pyyaml~=6.0.1
pytest~=8.2.2
requests~=2.32.3
sqlparse~=0.5.0
pydantic~=2.7.3
pydantic-core~=2.18.4
pandas~=2.2.2
flask~=3.0.3
flask-restful~=0.3.10
flask-swagger~=0.2.14
flask_swagger_ui~=4.11.1
flask-cors~=4.0.0
jinja2~=3.1.4
django~=5.0.6
sqlalchemy~=2.0.30
cerberus~=1.3.5
openpyxl~=3.1.2

